requirements.txt
pyaudio
argparse
SpeechRecognition
openai-whisper
torch